requirements.txt
gradio
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
accelerate>=0.20.0
tqdm>=4.65.0